pyproject.toml
[tool.poetry]
name = "csrf-backend"
version = "0.1.0"
description = ""
authors = ["Alexandr Alferov <aas@arenadata.io>"]
readme = "README.md"
packages = [{include = "csrf_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.6"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
